Refactor: move recent search to data.models
