+++ app/src/main/java/fr/free/nrw/commons/data/models/RecentSearch.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.explore.recentsearches
+package fr.free.nrw.commons.data.models
 
 import android.net.Uri
 import java.util.*
